+++ cli/cli/scripts/launch-cli.sh
@@ -16,7 +16,7 @@ source "${script_dirpath}/_constants.sh"
 # ==================================================================================================
 goarch="$(go env GOARCH)"
 goos="$(go env GOOS)"
-cli_binary_filepath="${cli_module_dirpath}/${GORELEASER_OUTPUT_DIRNAME}/${GORELEASER_CLI_BUILD_ID}_${goos}_${goarch}/${CLI_BINARY_FILENAME}"
+cli_binary_filepath="${cli_module_dirpath}/${GORELEASER_OUTPUT_DIRNAME}/${GORELEASER_CLI_BUILD_ID}_${goos}_${goarch}_v1/${CLI_BINARY_FILENAME}"
 if ! [ -f "${cli_binary_filepath}" ]; then
     echo "Error: Expected a CLI binary to have been built by Goreleaser at '${cli_binary_filepath}' but none exists" >&2
     exit 1