+++ cli/docs/changelog.md
@@ -1,6 +1,6 @@
 # TBD
 ### Breaking Changes
-* Upgraded to engine 1.26.0 and core 1.54.0
+* Upgraded to engine 1.26.1, core 1.54.1 and container-engine-lib 0.33.0
   * Users will need to run `kurtosis engine restart`
 
 ### Fixes